requirements.txt
# Core Dependencies
pandas>=2.0.0
sodapy>=2.2.0

# Google Cloud / BigQuery
google-auth>=2.0.0
pandas-gbq>=0.19.0
google-cloud-bigquery>=3.0.0

# Configuration Management
python-dotenv>=1.0.0
